add numeric

+++ database/migrations/2023_07_19_134811_create_products_table.php
@@ -22,8 +22,8 @@ class CreateProductsTable extends Migration
             $table->mediumText('small_description')->nullable();
             $table->longText('description')->nullable();
             $table->integer('original_price');
-            $table->integer('selling_price');
-            $table->integer('quantity');
+            $table->float('selling_price');
+            $table->float('quantity');
             $table->tinyInteger('trending')->default('0')->comment('1=trending,0=not-trending');
             $table->tinyInteger('featured')->default('0')->comment('1=featured,0=not-featured');
             $table->tinyInteger('status')->default('0')->comment('1=hidden,0=visible');
